Tests tidy up
