+++ src/SFA.DAS.Rofjaa.Api.UnitTests/Api/AddDatabaseExtensionTests.cs
@@ -1,4 +1,4 @@
-namespace SFA.DAS.Rofjaa.Api.Tests.AppStart
+namespace SFA.DAS.Rofjaa.Api.UnitTests.AppStart
 {
     using SFA.DAS.Rofjaa.Api.AppStart;
     using System;
@@ -10,14 +10,10 @@ namespace SFA.DAS.Rofjaa.Api.Tests.AppStart
     [TestFixture]
     public static class AddDatabaseExtensionTests
     {
-
-
         [Test]
         public static void CannotCallAddDatabaseRegistrationWithNullServices()
         {
             Assert.Throws<ArgumentNullException>(() => default(IServiceCollection).AddDatabaseRegistration(new RofjaaConfiguration { ConnectionString = "TestValue83455063" }, "TestValue1873433359"));
         }
-
-     
     }
 }
