+++ src/SFA.DAS.Rofjaa.Data.UnitTests/SFA.DAS.Rofjaa.Data.UnitTests.csproj
@@ -17,8 +17,12 @@
     </PackageReference>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.1" />
     <PackageReference Include="SFA.DAS.Testing.AutoFixture" Version="3.0.169" />
+    <PackageReference Include="NSubstitute" Version="4.3.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\SFA.DAS.Rofjaa.Data\SFA.DAS.Rofjaa.Data.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Remove="NSubstitute" />
   </ItemGroup>
 </Project>
