+++ src/SFA.DAS.Rofjaa.Domain.UnitTests/Courses/RofjaaConfigurationTests.cs
